backend/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
mcp>=1.0.0
uvicorn>=0.32.1
boto3
faiss-cpu>=1.7.4
sentence-transformers==2.2.2
numpy>=1.24.0
google-cloud-aiplatform>=1.39.0
google-auth>=2.0.0
opensearch-py>=2.3.1
azure-core>=1.26.0
azure-identity>=1.13.0
azure-search-documents>=11.4.0
fastapi
huggingface_hub==0.16.4
PyMuPDF>=1.22.5
